# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Socle SIRH — Sorbonne-Hôtel",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/etablissements": {
      "get": {
        "tags": [
          "Core HR"
        ],
        "summary": "Les cinq hôtels de la chaîne",
        "operationId": "lister_etablissements_etablissements_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EtablissementOut"
                  },
                  "type": "array",
                  "title": "Response Lister Etablissements Etablissements Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jeton de groupe": []
          }
        ]
      }
    },
    "/departements": {
      "get": {
        "tags": [
          "Core HR"
        ],
        "summary": "Les départements, filtrables par établissement",
        "operationId": "lister_departements_departements_get",
        "security": [
          {
            "Jeton de groupe": []
          }
        ],
        "parameters": [
          {
            "name": "etablissement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtre sur un hôtel",
              "title": "Etablissement Id"
            },
            "description": "Filtre sur un hôtel"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartementOut"
                  },
                  "title": "Response Lister Departements Departements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/postes": {
      "get": {
        "tags": [
          "Core HR"
        ],
        "summary": "Les postes, avec effectif cible et effectif réel",
        "operationId": "lister_postes_postes_get",
        "security": [
          {
            "Jeton de groupe": []
          }
        ],
        "parameters": [
          {
            "name": "etablissement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Etablissement Id"
            }
          },
          {
            "name": "departement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Departement Id"
            }
          },
          {
            "name": "famille",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Famille de métier, ex. « Cuisine »",
              "title": "Famille"
            },
            "description": "Famille de métier, ex. « Cuisine »"
          },
          {
            "name": "vacants",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "À true, ne renvoie que les postes dont l'effectif réel est inférieur à l'effectif cible",
              "title": "Vacants"
            },
            "description": "À true, ne renvoie que les postes dont l'effectif réel est inférieur à l'effectif cible"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PosteOut"
                  },
                  "title": "Response Lister Postes Postes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/postes/{poste_id}": {
      "get": {
        "tags": [
          "Core HR"
        ],
        "summary": "Un poste et les compétences qu'il requiert",
        "operationId": "lire_poste_postes__poste_id__get",
        "security": [
          {
            "Jeton de groupe": []
          }
        ],
        "parameters": [
          {
            "name": "poste_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Poste Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosteDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/competences": {
      "get": {
        "tags": [
          "Core HR"
        ],
        "summary": "Le référentiel de compétences",
        "operationId": "lister_competences_competences_get",
        "security": [
          {
            "Jeton de groupe": []
          }
        ],
        "parameters": [
          {
            "name": "categorie",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SkillCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categorie"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompetenceOut"
                  },
                  "title": "Response Lister Competences Competences Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/salaries": {
      "get": {
        "tags": [
          "Core HR"
        ],
        "summary": "Les salariés, avec leur contrat en cours",
        "operationId": "lister_salaries_salaries_get",
        "security": [
          {
            "Jeton de groupe": []
          }
        ],
        "parameters": [
          {
            "name": "etablissement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Etablissement Id"
            }
          },
          {
            "name": "departement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Departement Id"
            }
          },
          {
            "name": "poste_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poste Id"
            }
          },
          {
            "name": "manager_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "L'équipe directe d'un manager",
              "title": "Manager Id"
            },
            "description": "L'équipe directe d'un manager"
          },
          {
            "name": "statut",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EmployeeStatus"
                },
                {
                  "type": "null"
                }
              ],
              "default": "actif",
              "title": "Statut"
            }
          },
          {
            "name": "type_contrat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContractType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type Contrat"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Recherche sur le nom, le matricule ou l'e-mail",
              "title": "Q"
            },
            "description": "Recherche sur le nom, le matricule ou l'e-mail"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalarieOut"
                  },
                  "title": "Response Lister Salaries Salaries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/salaries/{salarie_id}": {
      "get": {
        "tags": [
          "Core HR"
        ],
        "summary": "Un salarié et son parcours contractuel",
        "operationId": "lire_salarie_salaries__salarie_id__get",
        "security": [
          {
            "Jeton de groupe": []
          }
        ],
        "parameters": [
          {
            "name": "salarie_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Salarie Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalarieDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DepartementOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "establishment_id": {
            "type": "integer",
            "title": "Establishment Id"
          },
          "establishment_name": {
            "type": "string",
            "title": "Establishment Name"
          },
          "establishment_city": {
            "type": "string",
            "title": "Establishment City"
          },
          "headcount": {
            "type": "integer",
            "title": "Headcount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "label",
          "establishment_id",
          "establishment_name",
          "establishment_city",
          "headcount"
        ],
        "title": "DepartementOut"
      },
      "EmployeeStatus": {
        "type": "string",
        "enum": [
          "actif",
          "sorti"
        ],
        "title": "EmployeeStatus"
      },
...
      "EtablissementOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "postal_code": {
            "type": "string",
            "title": "Postal Code"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "category": {
            "type": "integer",
            "title": "Category"
          },
          "opened_on": {
            "type": "string",
            "format": "date",
            "title": "Opened On"
          },
          "is_head_office": {
            "type": "boolean",
            "title": "Is Head Office"
          },
          "headcount": {
            "type": "integer",
            "title": "Headcount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "city",
          "postal_code",
          "address",
          "category",
          "opened_on",
          "is_head_office",
          "headcount"
        ],
        "title": "EtablissementOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PosteOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "job_family": {
            "type": "string",
            "title": "Job Family"
          },
          "hierarchy_level": {
            "type": "integer",
            "title": "Hierarchy Level"
          },
          "target_headcount": {
            "type": "integer",
            "title": "Target Headcount"
          },
          "current_headcount": {
            "type": "integer",
            "title": "Current Headcount"
          },
          "vacancies": {
            "type": "integer",
            "title": "Vacancies"
          },
          "department_id": {
            "type": "integer",
            "title": "Department Id"
          },
          "department_label": {
            "type": "string",
            "title": "Department Label"
          },
          "establishment_id": {
            "type": "integer",
            "title": "Establishment Id"
          },
          "establishment_name": {
            "type": "string",
            "title": "Establishment Name"
          },
          "establishment_city": {
            "type": "string",
            "title": "Establishment City"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "title",
          "job_family",
          "hierarchy_level",
          "target_headcount",
          "current_headcount",
          "vacancies",
          "department_id",
          "department_label",
          "establishment_id",
          "establishment_name",
          "establishment_city"
        ],
        "title": "PosteOut"
      },
...
      "SkillCategory": {
        "type": "string",
        "enum": [
          "technique",
          "relationnel",
          "langue",
          "management"
        ],
        "title": "SkillCategory"
      },
...
    },
    "securitySchemes": {
      "Jeton de groupe": {
        "type": "http",
        "description": "Le jeton remis à votre groupe, en en-tête `Authorization: Bearer <jeton>`.",
        "scheme": "bearer"
      }
    }
  }
}